LPD-96652 Move filters to a Filters component
