+++ modules/apps/site/site-cms-site-initializer/src/main/resources/META-INF/resources/js/main_view/dashboard/performance/PerformanceDashboard.tsx
@@ -3,26 +3,10 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-import ClayLayout from '@clayui/layout';
-import {
-	ConnectToAnalyticsCloud,
-	RangeSelector,
-	RangeSelectors,
-	RangeSelectorsDropdown,
-} from '@liferay/analytics-reports-js-components-web';
-import React, {useState} from 'react';
+import {ConnectToAnalyticsCloud} from '@liferay/analytics-reports-js-components-web';
+import React from 'react';
 
-import {
-	SpaceOption,
-	SpacesDropdown,
-	initialSpace,
-} from '../common/SpacesDropdown';
-
-const initialRangeSelector: RangeSelector = {
-	rangeEnd: '',
-	rangeKey: RangeSelectors.Last7Days,
-	rangeStart: '',
-};
+import {Filters} from './components/Filters';
 
 export default function PerformanceDashboard({
 	admin,
@@ -31,10 +15,6 @@ export default function PerformanceDashboard({
 	admin: boolean;
 	analyticsEnabled: boolean;
 }) {
-	const [rangeSelector, setRangeSelector] =
-		useState<RangeSelector>(initialRangeSelector);
-	const [space, setSpace] = useState<SpaceOption>(initialSpace);
-
 	if (!analyticsEnabled) {
 		return (
 			<div
@@ -46,30 +26,5 @@ export default function PerformanceDashboard({
 		);
 	}
 
-	return (
-		<ClayLayout.Row className="mb-4">
-			<ClayLayout.Col size={12}>
-				<div className="d-flex">
-					<SpacesDropdown
-						className="mr-3"
-						onSelectSpace={setSpace}
-						selectedSpace={space}
-					/>
-
-					<RangeSelectorsDropdown
-						activeRangeSelector={rangeSelector}
-						availableRangeKeys={[
-							RangeSelectors.Last24Hours,
-							RangeSelectors.Last7Days,
-							RangeSelectors.Last28Days,
-							RangeSelectors.Last30Days,
-							RangeSelectors.Last90Days,
-						]}
-						borderless={false}
-						onChange={setRangeSelector}
-					/>
-				</div>
-			</ClayLayout.Col>
-		</ClayLayout.Row>
-	);
+	return <Filters />;
 }
